refactor: split ImageStorageService into composed modules
The service mixed path resolution, image writing, directory management,
and orphaned-file scanning/deletion across 633 lines.

Split into a path resolver that owns the root directories, a writer, and
an orphan-cleanup module. ImageStorageService keeps all 14 public methods
and now composes those three (121 lines).

Two private members are deliberately retained as instance members:
uploads.service.ts and instagram.service.ts reach past `private` via
bracket indexing (imageStorage["baseImagesDir"],
imageStorage["cleanupEmptyFolders"]). Moving them wholesale into the
extracted modules left those four call sites resolving to undefined at
runtime — bun test does not cover them, so this surfaced only via tsc.

Also preserved: the two URL-fetch retry paths use different User-Agent
strings. Left as-is rather than unified, since collapsing them would
change what those requests send.

Verified: bun test 213 pass / 4 fail — identical to baseline; tsc error
list identical to baseline; runtime smoke test of the public surface plus
both bracket-accessed privates produces byte-identical output vs main.
